requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
sqlalchemy==2.0.45
python-jose[cryptography]==3.5.0
pydantic-settings==2.12.0
email-validator==2.3.0
bcrypt==5.0.0
python-multipart==0.0.21
passlib[bcrypt]==1.7.4
